vault: setup vlan and pppoe.
This changes allow to connect the Vault to Proximus' ONT. To make it
work, we need to use a vlan id 10, and run pppd.

There is further configuration for DHCP/SLAAC which I added when I
thought that what I needed was a vlan id 20 and no pppd (as Proximus
users with newer equipment do). That configuration is thus maybe
unnecessary and/or wrong.

A detail that was necessary to reach some websites: lower the MTU to
1492, both in pppd and hostapd.

That being said, with thoses changes, I can use the router as a wifi
access point and reach the internet.